Add locale utility functions to Utils class
- Add getUserLanguages() to detect browser language preferences
- Add findBestLocale() to match user preferences with available locales
- Export Utils class from index.js
- Update examples to use shared Utils functions
- Add Utils documentation with API reference
