+++ src/index.js
@@ -6,6 +6,7 @@ import LiturgyOfTheDay from "./LiturgyOfTheDay/LiturgyOfTheDay.js";
 import LiturgyOfAnyDay from "./LiturgyOfAnyDay/LiturgyOfAnyDay.js";
 import PathBuilder from "./PathBuilder/PathBuilder.js";
 import Input from './ApiOptions/Input/Input.js';
+import Utils from './Utils.js';
 import { Grouping, ColorAs, Column, ColumnOrder, DateFormat, GradeDisplay, ApiOptionsFilter, CalendarSelectFilter, YearType } from './Enums.js';
 
 export {
@@ -17,6 +18,7 @@ export {
     LiturgyOfAnyDay,
     PathBuilder,
     Input,
+    Utils,
     Grouping,
     ColorAs,
     Column,
